% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritizeCandidates.R
\name{prioritizeCandidates}
\alias{prioritizeCandidates}
\title{Candidate gene prioritization}
\usage{
prioritizeCandidates(counts, seedGenes, candidateGenes=NULL,
antiCorrelation=FALSE)
}
\arguments{
\item{counts}{A gene expression matrix (genes in rows, samples in columns):
can be a count matrix, a microarray result, as long as it's numeric. Rows
have to be named with the gene symbol or ID.
If desired, normalization has to be previously performed be the user.}

\item{seedGenes}{A vector or list of seed gene symbols or IDs. If a list,
each gene has to be a list element. At least one seed gene has to be present
in the counts matrix rownames. Missing or duplicated genes are removed.}

\item{candidateGenes}{(default NULL) A vector or list of candidate gene
symbols or IDs. If a list, each gene has to be a list element. If NULL,
all other genes in the matrix which are not part of the seed genes
will be considered candidates. If supplied, least one candidate gene
has to be present in the rank matrix rownames. Missing or duplicated genes
are removed.}

\item{antiCorrelation}{A logical value (default FALSE). If TRUE,
anti-correlation will be considered as a significant correlation, thus
strongly anti-correlated genes will have a high rank.}
}
\value{
A RankedGeneList object
}
\description{
Finds potential disease-related genes
based on their coexpression with known disease-related ("seed") genes.
Takes as input a gene expression matrix, a list of seed genes and
(optionally) a list of candidate genes; computes a ranked co-expression
list for each seed gene; assigns to each candidate gene the product of its
ranks within these lists as a total score; ranks them according to their
increasing score. Lower scores indicate a higher coexpression, and thus a
higher probability of being involved in the given phenotype. The function
performs all steps of the analysis and returns a \code{\link{RankedGeneList}}
object summarizing the
results, but they can be performed manually by the user with the functions
\code{\link{findCoexpression}}, \code{\link{rankGenes}}
and \code{\link{candidateScoring}}.
}
\examples{
a <- matrix(c(1,2,3,2,4,6,8,6,4,5,2,8,7,1,5),
nrow=5, ncol=3,byrow=TRUE)
colnames(a) <- c('sample1','sample2','sample3')
rownames(a) <- c('gene1','gene2','gene3','gene4','gene5')
seed <- c('gene1')
candidates <- c('gene2','gene4')
z <- prioritizeCandidates(counts=a, seedGenes=seed,
candidateGenes=candidates, antiCorrelation=FALSE)
}
\references{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2935433/}\cr
Piro, Rosario M et al. “Candidate gene prioritization based on spatially
mapped gene expression: an application to XLMR.” Bioinformatics (Oxford,
England) vol. 26,18 (2010): i618-24. doi:10.1093/bioinformatics/btq396
}
\seealso{
\code{\link{findCoexpression}}, \code{\link{rankGenes}},
\code{\link{candidateScoring}}, \code{\link{RankedGeneList}}\cr
}
\author{
Chiara Paleni\cr Politecnico di Milano\cr Maintainer: Chiara Paleni
\cr E-Mail: <chiara.paleni@polimi.it>
}
