% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGettersSetters.R
\name{RankedGeneList-accessors}
\alias{RankedGeneList-accessors}
\alias{getCandidates}
\alias{getSeed}
\alias{getResults}
\alias{getCoexpression}
\alias{getRanks}
\title{Accessor methods for class \code{\link{RankedGeneList}}}
\usage{
getCandidates(object)

getSeed(object)

getResults(object)

getCoexpression(object)

getRanks(object)
}
\arguments{
\item{object}{A RankedGeneList object}
}
\value{
A specific slot of the object
}
\description{
Accessor methods for class \code{\link{RankedGeneList}}
}
\examples{
a <- matrix(c(1,2,3,2,4,6,8,6,4,5,2,8,7,1,5),
nrow=5, ncol=3,byrow=TRUE)
colnames(a) <- c('sample1','sample2','sample3')
rownames(a) <- c('gene1','gene2','gene3','gene4','gene5')
seed <- c('gene1')
candidates <- c('gene2','gene4')
z <- prioritizeCandidates(counts=a, seedGenes=seed,
candidateGenes=candidates, antiCorrelation=FALSE)
getSeed(z)
}
\seealso{
\code{\link{RankedGeneList}},
\code{\link{prioritizeCandidates}}\cr
}
