% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{RankedGeneList-methods}
\alias{RankedGeneList-methods}
\alias{show,RankedGeneList-method}
\alias{summary,RankedGeneList-method}
\title{Methods for class RankedGeneList}
\usage{
\S4method{show}{RankedGeneList}(object)

\S4method{summary}{RankedGeneList}(object)
}
\arguments{
\item{object}{A RankedGeneList object}
}
\value{
Meaningful representations
}
\description{
Methods for class RankedGeneList
}
\seealso{
\code{\link{prioritizeCandidates}}, \code{\link{RankedGeneList}}\cr
}
\author{
Chiara Paleni\cr Politecnico di Milano\cr Maintainer: Chiara Paleni
\cr E-Mail: <chiara.paleni@polimi.it>
}
